This really huge set of changes reflects the time that has been spent on trying to understand and use the Jekyll web site generator tool.
These set of files contain files that will be altered in the future.
Though this is an extensive amount of changes.  There is a high
probability that most of the added files will be removed, replaced
and/or simplified.

The current changes will not be published to https://ccorsi.github.io
until another set of changes will be applied.  The original learning
web site will be generated through this repository instead of the
learning repository.  Ulimately, that web site will be integrated into a
more general site that will have a better transversal of the content.

Finally, this set of changes produces several different web sites that
will be generated within a sub-directory from the top-level web site
url.  These are currently:

	- webone
	- webtwo
	- webthree
	- webfour
	- learning
	- experimental
	- documentation
	- testdefault
	- testblank