The following set of changes is the first pass at trying to build a site that will can be used to create more elaborate pages that are better than my initial try at using the Jekyll tool with the Beautiful Jekyll theme.  The original attempt is part of the learning/ link on the generated site.
While the current set of changes include many more changes and files
that will ultimately not be part of the generated site.  This was on
purpose more as a historical reason than anything else.  These files
will eventually be removed from the repository afte documenting my
experience using those generated sites.

Finally, I have replaced the current site build with a more simplier
version that will generate the main site as well as all the other "test"
sites that I have used to better understand how to use Jekyll and the
different themes that I have used along the way.