This set of changes includes the removal of the docker run command since it isn't required any longer to be able to build and publish the site.
The table of content generation now assigns a row class instance of a
container-md class.  This makes the generation of the table of content
to be view in a better way.

The layout of the sidebar and footnote page to be using the default
layout instead of the page layout.  This will provide more viewing space
in the page so that the generated page will be less compact.