This set of changes includes the moving of the favicon settings from the assets/images directory to the assets/img directory to keep it simplier.
Created the projects and programming tools sidebar definition.  This is
to initially create the required list of files that will be brought over
from the original learning repository to this repository.

Created a layout that extends the default layout from the beautiful
jekyll theme.  This was done because I didn't really like the page
layout as much so this one was introduced to include my own take on it.

Created the main projects page and imported the original dataloader page
to this repository.  This work included fixing many grammar issues and
reworked somewhat the wording of the page.  As a whole the prior page
was not changed more.

Updated the About page somewhat and added an initial table css styles
that can be used throughout in the future.

Finally, updated the scope entries in the configuration file to include
quotes and a table of content for each of the different set of pages.