This follow set of changes includes the moving of the math reference to page to its own top-level references directory.  This was done to be able to define reference specific front matter settings within the _config.yml file.
Created javascript code that is used by the generated reference file to
determine if the page was accessed through a link from one of the math
pages.  This code will then update the page such that you can return to
the prior page by just clicking the highlighted reference entry.

Added another reference to the list of mathematical reference file.
This is being referenced by the updated number theory pages that were
edited within these set of changes.

Created a citation include file that is used to generate the correct
citation used within a page.  The citation include will use the passed
values to create the appropiate link to the reference page.