Updated and added sidebars and subbars associated with the Elementary number theory section.  This is part of initially adding an integers section of the elementary number theory section.
Added a proposition include that will be used whenever mathematical
propositions are defined within a page.

Updated the definition include to handle some issues that I noticed
along the way of using this include file.  These changes will be added
to the other mathematical include files that use the same set of
variables to generate the different mathematical statements.

Created a mathjax-config include file to allow for user defined mathjax
settings.

Updated the math.scss file to include definitions for the lemma,
proposition and corollary.

Updated the math-footnotes.js file such that it uses the definition
within the front matter to setup which types of mathematical statements
should have footnotes created for them.

Finally, added an elementary number theory integer section page.