The following set of changes includes the layout definition for posts, notes and compsc sections of the web site.  While the posts section is currently being used.  The notes and compsc sections are currently not being populated or published.
Updated the create.sh script to be able to create an initial post file.

While the run.sh script was added the -t command line option that states
that only the top-level _site site should be generated.

Unpublished the original post that was generated with the use of the
beautiful jekyll theme and replaced it with my "Here We Go Again" post.