Added the following books to the reading page:
   - The Joy of X
   - Taming Silicon Valley
   - Nexus

While the book "The Memory Police" was moved to the read page.

Updated the config file to exclude the about.htm file since this is not
the page that was supposed to be used as my about page.  Thou, this page
is being overriden by the about.md page.

Added 3 more quotes to the quotes data file.